+++ src/sections/Tickets/components/index.js
@@ -1,2 +1,2 @@
-export { default as MainModal } from './MainModal.vue';
-export { default as MainTable } from './MainTable.vue';
+export { default as MainModal } from './TicketModal.vue';
+export { default as MainTable } from './TicketTable.vue';
